Create README.md
Migrant Stock UN definition.

Dataset

According to UN.org site Migran Stock is defined as ¨the estimates of the number (or "stock") of international migrants disaggregated by age, sex and country or area of origin are based on national statistics, in most cases obtained from population censuses. Additionally, population registers and nationally representative surveys provided information on the number and composition of international migrants¨.

The dataset used presents estimates of international migrant by age, sex and origin. Estimates are presented for 1990, 1995, 2000, 2005, 2010, 2015 and 2020 and are available for 232 countries and areas of the world. The estimates are based on official statistics on the foreign-born or the foreign population.

Scope of the Analysis

Our analysis are based only on data corresponding to Panama, central american country which this study is focused.

Purpose of the Analysis

Identify the behavior of migratory flows from different regions to Panama and its evolution from the year 1990 to 2020.